requirements.txt
# hybrid-llm-inference/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyjoules>=0.5.1
transformers>=4.30.0
accelerate>=0.20.0
requests>=2.28.0
pytest>=7.3.0
pytest-cov>=4.0.0
pyyaml>=6.0.0
psutil>=5.9.0
